refactor(): moved helpers vdom func
