+++ src/core/__tests__/vdom.spec.ts
@@ -1,7 +1,7 @@
 /**
  * @jest-environment jsdom
  */
-import { h, changed, render, setProps, setProp, setBooleanProp } from "../vdom";
+import { h, changed, render } from "../vdom";
 
 describe("h function", () => {
   test("it renders the h function with required arguments", () => {
@@ -87,59 +87,13 @@ describe("Update element", () => {
   });
 });
 
-describe("internal functions", () => {
-  test("changed function", () => {
+describe("changed functions", () => {
+  test("it should have changed", () => {
     const node1 = h("div", { id: "one" }, "Node 1");
     const node2 = h("button", { id: "two" }, "Node 2");
 
     expect(changed(node1, node1)).toEqual(false);
     expect(changed(node1, node2)).toEqual(true);
     expect(changed(render("div"), node1)).toEqual(true);
-  });
-
-  test("setProps", () => {
-    const props = {
-      id: "foo",
-      className: "some-classname",
-      title: "some title",
-    };
-    const anchor = document.createElement("a");
-    setProps(anchor, props);
-
-    expect(anchor.id).toBe("foo");
-    expect(anchor.title).toBe("some title");
-    expect(anchor.className).toBe("some-classname");
-  });
-
-  test("setProp with classname", () => {
-    const target = document.createElement("div");
-    setProp({ target, name: "className", value: "a-classname" });
-    expect(target.className).toBe("a-classname");
-  });
-
-  test("setProp with a data-title", () => {
-    const target = document.createElement("div");
-    setProp({ target, name: "id", value: "data-title" });
-    expect(target.id).toBe("data-title");
-  });
-
-  test("setProp with boolean", () => {
-    const target = document.createElement("div");
-    setProp({ target, name: "boolean", value: "true" });
-    expect(target.attributes.item(0)?.name).toEqual("boolean");
-    expect(target.attributes.item(0)?.value).toEqual("true");
-  });
-
-  test("setBooleanProp on true", () => {
-    const target = document.createElement("div");
-    setBooleanProp({ target, name: "boolean", value: "true" });
-    expect(target.attributes.item(0)?.name).toEqual("boolean");
-    expect(target.attributes.item(0)?.value).toEqual("true");
-  });
-
-  test("setBooleanProp on false", () => {
-    const target = document.createElement("div") as any; // use a better type or different kind of test
-    setBooleanProp({ target, name: "boolean", value: "" });
-    expect(target.boolean).toEqual(false);
   });
 });
